+++ src/lib/bars/chart-bar-grouped.ts
@@ -1,4 +1,5 @@
 import { BaseType, select, Selection } from 'd3-selection';
+import { axisTickFindByIndex, axisTickHighlight } from '../axis';
 import { COLORS_CATEGORICAL, debug, nodeToString } from '../core';
 import {
   chartCartesian,
@@ -6,10 +7,22 @@ import {
   dataChartCartesian,
   DataChartCartesian,
 } from '../core/chart-cartesian';
-import { DataLegendSquares, dataLegendSquares, legend } from '../legend';
-import { DataBar, JoinEvent, seriesBar } from './series-bar';
-import { DataSeriesBarGrouped, dataSeriesBarGrouped, seriesBarGrouped } from './series-bar-grouped';
-import { seriesLabel } from './series-label';
+import {
+  DataLegendSquares,
+  dataLegendSquares,
+  legend,
+  legendItemFindByIndex,
+  legendItemFindByLabel,
+  legendItemHighlight,
+} from '../legend';
+import { barFind, DataBar, JoinEvent, seriesBar } from './series-bar';
+import {
+  DataBarGrouped,
+  DataSeriesBarGrouped,
+  dataSeriesBarGrouped,
+  seriesBarGrouped,
+} from './series-bar-grouped';
+import { labelFind, labelHighlight, seriesLabel } from './series-label';
 import { dataSeriesLabelBar } from './series-label-bar';
 
 export interface DataChartBarGrouped extends DataSeriesBarGrouped, DataChartCartesian {
@@ -49,6 +62,9 @@ export function chartBarGrouped<
         .call((s) => seriesBarGrouped(s))
         .on('barupdate', (e: JoinEvent<SVGRectElement, DataBar>) =>
           e.detail.selection.attr('fill', (d) => chartData.colors[d.index])
+        )
+        .on('mouseover.chartbargroupedhighlight mouseout.chartbargroupedhighlight', (e) =>
+          chartBarGroupedHoverBar(chart, select(e.target), e.type.endsWith('over'))
         );
 
       drawArea
@@ -96,4 +112,20 @@ export function chartBarGroupedDataChange<
     chartData.crossAxis.scale = chartData.crossScale;
     chartCartesianUpdateAxes(s);
   });
+}
+
+export function chartBarGroupedHoverBar(
+  chart: Selection,
+  bar: Selection<Element, DataBarGrouped> | string,
+  hover: boolean
+) {
+  const barS = typeof bar === 'string' ? barFind<DataBarGrouped>(chart, bar) : bar,
+    barD = barS.datum(),
+    labelS = labelFind(chart, barD.key),
+    tickS = axisTickFindByIndex(chart.selectAll('.axis-main'), barD.groupIndex),
+    legendItemS = legendItemFindByIndex(chart, barD.index);
+
+  labelHighlight(labelS, hover);
+  axisTickHighlight(tickS, hover);
+  legendItemHighlight(legendItemS, hover);
 }